requirements.txt
Flask==3.1.0
requests~=2.32.3
youtube_transcript_api~=0.6.2
urllib3~=2.2.3
python-dotenv==1.0.0
